requirements.txt
# Web Interface
streamlit>=1.55.0

# LLM & RAG Framework
llama-index-llms-google-genai>=0.6.2
llama-index-embeddings-google-genai>=0.1.0
llama-index>=0.12.0
llama-index-vector-stores-chroma>=0.5.5

# Document Parsing (High-quality PDF reading)
llama-parse>=0.5.0

# Vector Database (Local storage)
chromadb>=0.6.0

# Environment Variables (For your API Key)
python-dotenv>=1.2.2